+++ tests/stubs/Queries/Ping.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Butler\Graphql\Tests\Queries;
+
+class Ping
+{
+    public function __invoke($root, $args, $context)
+    {
+        return "pong";
+    }
+}