+++ CHANGELOG.md
@@ -7,14 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Decorate the response with debug information when using [barryvdh/laravel-debugbar](https://github.com/barryvdh/laravel-debugbar).
+
 ### Changed
 - Upgrade webonyx/graphql-php@0.12.6 for improved schema language support.
 
 ### Fixed
 - The code in `HandlesGraphqlRequests->errorFormatter()` always assumed `\Exception` from `\GraphQL\Error\Error->getPrevious()` when it fact it returns `\Throwable`. We now ensure that we always pass `\Exception` to `ExceptionHandler->report()`.
 
-### Added
-- Add support for barryvdh/laravel-debugbar.
 
 ## [1.1.2] - 2018-09-07